+++ ui/editor/key/undo-redo.ts
@@ -8,8 +8,6 @@ export class Undo extends KeyHandler {
 
   handle(): void {
     this.editor.history.undo();
-
-    this.editor.on_has_changes?.(this.editor.history.has_changes);
   }
 }
 
@@ -21,7 +19,5 @@ export class Redo extends KeyHandler {
 
   handle(): void {
     this.editor.history.redo();
-
-    this.editor.on_has_changes?.(this.editor.history.has_changes);
   }
 }